Adding parentize method, to which adds an .parent() method to a prime.
parentize or .parent(method) adds the parent method to the current prime
prototype, as "_parent_method", and .parent('method') then calls
this._parent_method().

This makes the .parent method considerably easier and probably faster.
It doesn't have to store the methods' result or delete the .parent method
that could cause troubles with multiple layers of inheritance.